services/web/requirements.txt
attrs==23.2.0
bcrypt==4.0.1
blinker==1.6.2
bokeh==3.2.1
certifi==2024.2.2
cffi==1.15.1
click==8.1.4
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cryptography==41.0.3
fiona==1.9.6
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
geopandas==0.14.3
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.8.0
isc-py-common==0.1.27
iscpy==1.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
Pillow==10.0.0
protobuf==3.20.3
pycparser==2.21
pymssql==2.2.7
PyMySQL==1.1.0
PyNaCl==1.5.0
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
shapely==2.0.3
six==1.16.0
SQLAlchemy==2.0.18
sshtunnel==0.4.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.6
WTForms==3.0.1
xyzservices==2023.7.0
zipp==3.16.2
